% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data_file.R
\name{build_data_file}
\alias{build_data_file}
\title{build_data_file}
\usage{
build_data_file()
}
\value{
Returns and/or stores and object called data_file which contains
a list of data for all requested electoral contexts.
}
\description{
Runs a ShinyApp that guides users through the steps
required for data download, recoding, mapping, and imputation.
}
