% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meas_error.R
\name{calculate_meas_error}
\alias{calculate_meas_error}
\title{Calculate measurement error in vote switching survey data}
\usage{
calculate_meas_error(
  switches,
  mappings = NULL,
  switch_factor = NULL,
  assign_others = NULL,
  type = c("mae", "rmse")
)
}
\arguments{
\item{switches}{A switching object produced by \code{build_data_file}.}

\item{mappings}{Optional: A copy of \link{mappings}, including a
column \link{switch_factor}.}

\item{switch_factor}{Optional: Column name of a factor variable in \code{mappings}
that specifies the marginal categories of the desired voter transition
matrix. If not supplied, the errors will be calculated on the original
transition matrix will nominal parties included in \code{mappings},
augmented with residual others and non-voters.}

\item{assign_others}{Optional: A character that specifies a level for
residual other parties for which \code{switch_factor} information is not
available.}

\item{type}{Specifies whether election-level mean absolute errors
or root mean squared errors will be returned. Defaults to \code{"mae"}.}
}
\value{
Returns a list with observation-/outcome-specific absolute errors and
the overall mean absolute error (if \code{type == "mae"}) or the overall
root mean squared error (if \code{type == "rmse"})

Returns a list of party-election level directional errors and
election-level mean absolute errors or root mean squared errors.
}
\description{
Uses an object of class \code{mavcl_est} and calculates
prediction errors of posterior means or medians.
}
