% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pred_error.R
\name{calculate_pred_error}
\alias{calculate_pred_error}
\title{calculate_pred_error}
\usage{
calculate_pred_error(
  mavcl_object,
  posterior = c("mean", "median"),
  type = c("mae", "rmse"),
  re_null = FALSE
)
}
\arguments{
\item{mavcl_object}{An object of class \code{mavcl_estimation}.}

\item{posterior}{\code{"mean"} or \code{"median"}, depending on whether posterior
means or medians should be evaluate against the observed proportions of the
outcome variable.}

\item{type}{\code{"mae"} or \code{"rmse"}, depending on whether (mean) average
errors or the root mean squared standard error should be reported}

\item{re_null}{logical, specifies whether the prediction should be calculated
with (\code{FALSE}) or without (\code{TRUE}) random effects}
}
\value{
Returns a list with observation-/outcome-specific absolute errors and
the overall mean absolute error (if \code{type == "mae"}) or the overall
root mean squared error (if \code{type == "rmse"})
}
\description{
Uses an object of class \code{mavcl_estimation} and calculates
prediction errors of posterior means or medians.
}
