% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\docType{data}
\name{codebook}
\alias{codebook}
\title{Documentations of mappings}
\format{
A data frame with 59 rows and 5 columns.
}
\usage{
codebook
}
\description{
A detailed documentation of \code{voteswitchR}'s mapping file, detailing
variables, their sources, and levels (country, election, party).
}
\keyword{datasets}
