% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_guide.R
\docType{data}
\name{data_guide}
\alias{data_guide}
\title{Versions, download links, and access details for survey data}
\format{
A data frame with 220 rows and 7 columns.
}
\usage{
data_guide
}
\description{
A detailed documentation of the survey data that \code{voteswitchR}'s
\code{\link{build_data_file}} function processes.
}
\keyword{datasets}
