% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_switches.R
\name{recode_switches}
\alias{recode_switches}
\title{Recode aggregate switching patterns}
\usage{
recode_switches(
  switches,
  mappings = NULL,
  switch_factor,
  type = c("elections", "party-elections")
)
}
\value{
Returns the supplied data frame \code{mappings}, extended by columns
for each switching pattern detected in the data.
}
\description{
Recodes raw aggregate switching patterns according to a
user-supplied factor variable included in a (possibly user-supplied version
of) \code{mappings}.
}
