% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteswitchR-package.R
\docType{package}
\name{voteswitchR-package}
\alias{voteswitchR-package}
\alias{voteswitchR}
\title{Data and Methods for Analyzing Comparative Vote Switching Data}
\description{
The \code{voteswitchR} package is...
}
\details{
The 'voteswitchR' package.
}
\section{Functions}{


\code{voteswichR} (development version 0.1.0) consists of ... functions:
\enumerate{
\item \code{...}: ...
}
}

\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
