% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_qoi.R
\name{compute_qoi}
\alias{compute_qoi}
\title{Compute MAVCL quantities of interest}
\usage{
compute_qoi(
  mavcl_object,
  y_structure,
  posterior_quantiles = c(0.5, 0.025, 0.975),
  len_continuous_sequence = 21L,
  conditional_expectation = TRUE,
  average_marginal_effect = TRUE,
  ame_shift = "tiny",
  base = c("t", "tm1", "avg"),
  relative = TRUE,
  atmeans = FALSE,
  re_null = FALSE
)
}
\value{
Returns a nested list of class \code{voteswitchR_qoi} that contains
the requested quantities.
}
\description{
Computes quantities of interest such as conditional expectations
and average marginal effects/first differences from a MAVCL estimation
object.
}
