% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\docType{data}
\name{mappings}
\alias{mappings}
\title{Mapping of vote choices from surveys to ParlGov and MARPOR IDs}
\format{
A data frame with 1547 rows and 59 columns. For details and sources
of each column, see \code{\link{codebook}}.
}
\usage{
mappings
}
\description{
\code{voteswitchR}'s mapping file, linking party-specific survey items
to ParlGov and MARPOR IDs for over 1,500 parties
}
\keyword{datasets}
