% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_switches.R
\name{recode_switches}
\alias{recode_switches}
\title{Recode aggregate switching patterns}
\usage{
recode_switches(
  switches,
  subgroup = NULL,
  stack_by_subgroup = FALSE,
  assign_others = NULL,
  mappings = NULL,
  switch_factor,
  type = c("elections", "party-elections")
)
}
\arguments{
\item{switches}{Aggregate counts of switches as generated by
\code{\link[=build_data_file]{build_data_file()}}.}

\item{subgroup}{A character that specifies the column-name of electorate
subgroups. Should be \code{NULL} (the default) if no subgroup analyses are
required.}

\item{stack_by_subgroup}{Logical; if \code{TRUE}, subgroups are stacked into
long format such that switching proportions from the MAVCL model sum to one
within subgroups.
If \code{FALSE}, subgroup-specific switching is recorded in wide format such
that switching proportions for each subgroup sum to the size of the subgroup.
Only takes effect if \code{subgroup != NULL}.}

\item{assign_others}{Optional: A character that specifies a level for
residual other parties for which \code{switch_factor} information is not
available.}

\item{mappings}{A copy of \link{mappings}, possibly
enriched with external party and/or election-specific variables.}

\item{switch_factor}{Column name of a factor variable in \code{mappings}
that specifies the marginal categories of the desired voter transition
matrix.}

\item{type}{Specifies whether to return election-specific voter-transition
matrices (\code{"elections"}) or stacked party-within-election transition
matrices (\code{"party-elections"}). The latter
should be used for studying the effects of party strategies.}
}
\value{
Returns the supplied data frame \code{mappings}, extended by columns
for each switching pattern detected in the data.
}
\description{
Recodes raw aggregate switching patterns according to a
user-supplied factor variable included in a (possibly user-supplied) version
of \code{mappings}.
}
