% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_switches.R
\name{recode_switches}
\alias{recode_switches}
\title{Recode aggregate switching patterns}
\usage{
recode_switches(
  switches,
  mappings = NULL,
  switch_factor,
  type = c("elections", "party-elections")
)
}
\arguments{
\item{switches}{Aggregate counts of switches as generated by
\code{\link{voteswitchR::build_data_file()}}.}

\item{mappings}{A copy of \code{\link{voteswitchR::mappings}}, possibly
enriched with external party and/or election-specific variables.}

\item{switch_factor}{Column name of a factor variable in \code{mappings}
that specifies the marginal categories of the desired voter transition matrix.}

\item{type}{Specifies whether to return election-specific voter-transition
matrices (\code{"elections"}) or stacked party-within-election transition
matrices (\code{"party-elections"}). The latter
should be used for studying the effects of party strategies.}
}
\value{
Returns the supplied data frame \code{mappings}, extended by columns
for each switching pattern detected in the data.
}
\description{
Recodes raw aggregate switching patterns according to a
user-supplied factor variable included in a (possibly user-supplied version
of) \code{mappings}.
}
