% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mavcl.R
\name{run_mavcl}
\alias{run_mavcl}
\title{run_mavcl}
\usage{
run_mavcl(
  data,
  Y_names,
  null_model = FALSE,
  main_predictor,
  predictor_continuous = TRUE,
  moderator = NULL,
  moderator_continuous = TRUE,
  other_covariates = NULL,
  random_slopes = FALSE,
  re_parties = FALSE,
  re_elections = FALSE,
  re_countries = FALSE,
  n_iter = 2000L,
  n_warm = 1000L,
  n_thin = 2L,
  n_chains = 2L,
  n_cores = parallel::detectCores(),
  max_treedepth = 15,
  adapt_delta = 0.8,
  seed,
  parallelize = FALSE,
  savename = "mavcl",
  savepath = NULL,
  outfile = TRUE
)
}
\value{
Returns an object of class \code{mavcl_estimation}: A list that
includes estimates and auxiliary information.
}
\description{
Runs the Mixed Aggregate Varying Choice Set Logit (MAVCL)
model via \code{rstan}.
}
