% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteswitchR-package.R
\docType{package}
\name{voteswitchR-package}
\alias{voteswitchR-package}
\alias{_PACKAGE}
\alias{voteswitchR}
\title{Data and Methods for Analyzing Comparative Vote Switching Data}
\description{
The \code{voteswitchR} package offers functions for data
management and statistical analyses to implement the conceptual
framework presented in Cohen, Krause and Abou-Chadi (2021).

\code{voteswichR} (development version 0.1.0) consists of the following
functions:

For data processing:
\enumerate{
\item \code{\link{build_data_file}}: Harmonize, impute, map, rake, and
aggregate vote switching data
\item \code{\link{recode_switches}}: Recode aggregate switching patterns
}

For model estimation and post-estimation:
\enumerate{
\item \code{\link{run_mavcl}}: Run the MAVCL model
\item \code{\link{calculate_pred_error}}: Calculate the MAEs and RMSEs of
MAVCL Predictions
\item \code{\link{compute_qoi}}: Compute MAVCL quantities of interest
}

Package-specific data:
\enumerate{
\item \code{\link{mappings}}: Mapping of vote choices from surveys to
ParlGov and MARPOR IDs.
\item \code{\link{codebook}}: Documentations of \code{\link{mappings}}.
\item \code{\link{data_guide}}: Versions, download links, and access
details for survey data.
}
}
\details{
The 'voteswitchR' package.
}
\references{
\enumerate{
\item Cohen, Denis, Krause, Werner and Abou-Chadi, Tarik (2022).
Comparative vote switching: A new framework for
studying dynamic multiparty competition. Conference Paper. Unpublished
manuscript.
\item Stan Development Team (2020). RStan: the R interface to Stan. R
package version 2.21.2. https://mc-stan.org
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/denis-cohen/voteswitchR}
  \item Report bugs at \url{https://github.com/denis-cohen/voteswitchR/issues}
}

}
\author{
\strong{Maintainer}: Denis Cohen \email{denis.cohen@mzes.uni-mannheim.de}

Other contributors:
\itemize{
  \item Andreas Küpfer \email{akuepfer@mail.uni-mannheim.de} [contributor]
  \item Tim Allinger \email{tialling@mail.uni-mannheim.de} [contributor]
  \item Nick Baumann \email{nbaumann@mail.uni-mannheim.de} [contributor]
  \item Werner Krause \email{werner.krause@hu-berlin.de} [contributor]
}

}
\keyword{internal}
