% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteswitchR-package.R
\docType{package}
\name{voteswitchR-package}
\alias{voteswitchR-package}
\alias{_PACKAGE}
\alias{voteswitchR}
\title{Data and Methods for Analyzing Comparative Vote Switching Data}
\description{
The \code{voteswitchR} package offers functions for data
management and statistical analyses to implement the conceptual
framework presented in Cohen, Krause and Abou-Chadi (2021).
}
\details{
The 'voteswitchR' package.
}
\references{
Cohen, Denis, Krause, Werner and Abou-Chadi, Tarik (2021). Getting the most out of comparative vote-switching data: A new framework for studying dynamic multiparty competition. Conference Paper. Unpublished manuscript.
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/denis-cohen/voteswitchR}
  \item Report bugs at \url{https://github.com/denis-cohen/voteswitchR/issues}
}

}
\author{
\strong{Maintainer}: Denis Cohen \email{denis.cohen@mzes.uni-mannheim.de}

Other contributors:
\itemize{
  \item Andreas Küpfer \email{akuepfer@mail.uni-mannheim.de} [contributor]
  \item Tim Allinger \email{tialling@mail.uni-mannheim.de} [contributor]
  \item Nick Baumann \email{nbaumann@mail.uni-mannheim.de} [contributor]
  \item Werner Krause \email{werner.krause@hu-berlin.de} [contributor]
}

}
\keyword{internal}
