% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteswitchR-package.R
\docType{package}
\name{voteswitchR-package}
\alias{voteswitchR-package}
\alias{voteswitchR}
\title{Data and Methods for Analyzing Comparative Vote Switching Data}
\description{
The voteswitchR package is an open-source R package that offers
a suite of functions for processing and analyzing vote switching data. It
offers a comprehensive software implementation of the conceptual framework
and the various extensions presented in Cohen, Krause, and Abou-Chadi
(2024).

\code{voteswichR} consists of the following
functions:

For data processing:
\enumerate{
\item \code{\link{build_data_file}}: Harmonize, impute, map, rake, and
aggregate vote switching data
\item \code{\link{recode_switches}}: Recode aggregate switching patterns
\item \code{\link{calculate_meas_error}}: Calculate measurement error in
vote switching survey data
}

For model estimation and post-estimation:
\enumerate{
\item \code{\link{run_mavcl}}: Run the MAVCL model
\item \code{\link{calculate_pred_error}}: Calculate the MAEs and RMSEs of
MAVCL Predictions
\item \code{\link{compute_qoi}}: Compute MAVCL quantities of interest
}

Package-specific data:
\enumerate{
\item \code{\link{mappings}}: Mapping of vote choices from surveys to
ParlGov and MARPOR IDs.
\item \code{\link{codebook}}: Documentations of \code{\link{mappings}}.
\item \code{\link{data_guide}}: Versions, download links, and access
details for survey data.
\item \code{\link{switches}}: Cell counts for the raw, idiosyncratic transition matrices for all elections included in \code{voteswitchR}.
\item \code{\link{switches_imp}}: 25 imputations of the cell counts for the raw, idiosyncratic transition matrices for all elections included in \code{voteswitchR}.
\item \code{\link{raked_switches}}: Rake-weighted cell counts for the raw, idiosyncratic transition matrices for all elections included in \code{voteswitchR}.
\item \code{\link{raked_switches_imp}}: 25 imputations of the rake-weighted cell counts for the raw, idiosyncratic transition matrices for all elections included in \code{voteswitchR}.
}
}
\details{
The 'voteswitchR' package.
}
\references{
\enumerate{
\item Cohen, Denis, Krause, Werner and Abou-Chadi, Tarik (2024).
Comparative vote switching: A new framework for
studying dynamic multiparty competition. The Journal of Politics.
\item Stan Development Team (2024). RStan: the R interface to Stan. R
package version 2.32.6. https://mc-stan.org
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/denis-cohen/voteswitchR}
  \item Report bugs at \url{https://github.com/denis-cohen/voteswitchR/issues}
}

}
\author{
\strong{Maintainer}: Denis Cohen \email{denis.cohen@mzes.uni-mannheim.de}

Other contributors:
\itemize{
  \item Andreas Küpfer \email{akuepfer@mail.uni-mannheim.de} [contributor]
  \item Tim Allinger \email{tialling@mail.uni-mannheim.de} [contributor]
  \item Nick Baumann \email{nbaumann@mail.uni-mannheim.de} [contributor]
  \item Werner Krause \email{werner.krause@hu-berlin.de} [contributor]
}

}
\keyword{internal}
