% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteswitchR.R
\docType{package}
\name{voteswitchR}
\alias{voteswitchR}
\title{voteswitchR: Build, Map, and Analyze Vote Switching Data in R}
\description{
The \code{voteswitchR} package is...
}
\section{Functions}{


\code{voteswichR} (development version 0.1.0) consists of ... functions:
\enumerate{
\item \code{...}: ...
}
}

\seealso{
\code{\link{...}}
}
